+++ Sources/VComponents/Components/Modals/SideBar/VSideBarUIModel.swift
@@ -143,7 +143,7 @@ public struct VSideBarUIModel {
 
     // MARK: Dismiss Type
     /// Dismiss type, such as `backTap, or `dragBack`.
-    @OptionSetRepresentation<UInt64>(accessLevelModifier: "public")
+    @OptionSetRepresentation<Int>(accessLevelModifier: "public")
     public struct DismissType {
         // MARK: Options
         private enum Options: Int {
